requirements.txt
PySide6>=6.5.0
Pillow>=10.0.0

